mpris: prefer instance names when --all is used
When a player has both a 'normal' name and an 'instance' name, e.g. VLC
("vlc" and "vlc.instance1234"), apply the action only to instance names,
to avoid having the action be applied twice to the oldest instance
(which is currently holding the 'normal' name).

`mpris --list` still returns the complete list of names, to avoid hiding
potentially useful information.