ctl: add a weird syntax for using assignments and shell globs
The shell won't expand globs in assignments because it thinks the value
is part of file name, so this won't work:

    ctl /proc/sys/net/ipv4/conf/*/promote_secondaries=1

To make it work, the assignment can now be split into a separate
argument. The simplest implementation was to make it set the 'default'
value for *future* assignments, using a syntax like this:

    ctl =1 /proc/sys/net/ipv4/conf/*/promote_secondaries

Looks weird, though.