timestamp: nuke
It's a mere four-liner and it still took me a while to understand its
purpose. Apparently it sneaked in along with kprop.ssh – running
`timestamp foo.db kprop` only runs kprop if 'foo.db' is newer than the
last-run timestamp 'foo.db.time', and – screw it, I'll just write it
again if I _ever_ need it.