authorized_keys.py: update comment
(Apparently people actually _use_ my code – I've found parts of it in at
least three different projects on Github. Which is *cool*. Apparently, also,
people don't like to attribute the code author. Which is less cool. But, I'm
not really worried about that, it's just a few lines of code, and it's under
the WTFPL for a reason – to make it easier for people to use my code.

But the thing is, if your copy has an URL to the original (even without the
author's name), then future-you and future-other-developers will know where
to check for *updates*, improvements, etc. For example, the "future-algo"
code I've added recently. Perhaps bugfixes. If you only copy-paste the code
without keeping any backlinks, then your copy has to evolve all by itself.)