Revert "scrape-reddit-comment: all hail .json"
The JSON makes it fairly difficult to distinguish between self-post
replies and comment replies. That is, a .json of a self-post will
include the post itself & the first comment. But the .json of any other
comment will include the post itself & the requested comment, so
everything is in the same format.

This reverts commit 9ca4ed1af3b5ed431e80471853d5a76c0b9732d4.