add `what` (find package versions in repositories)
This is a bit like `pacman -Q` but for packages available in
repositories. (Unfortunately `pacman -Si` doesn't have a "terse"
mode...) In the near future, this might be replaced with

    expac -S "%r/%n %v" "$@"

which right now cannot deal with repositories that are Included into the
main pacman.conf, so unfortunately skips a few packages.