secretsd: remove Item, rely exclusively on fallback
The original plan was to have everything as a full object, but that
seems unnecessarily expensive, in terms of both performance and
complexity. For example, Item objects don't actually have much of a
state on their own – everything can be just retrieved from SQLite on
demand. (Though if *that* ends up slowing things down more, I'll
reconsider...)