annex purge: add get/put
Ugh.

Ugh.

I've found myself using the 'unused' feature as a sort of Recycle Bin
for my animes – deleting them from the file tree but keeping the actual
blobs in storage (eyes don't see, heart doesn't feel) – it's even gotten
to the point where I don't want to purge unused files because "oh maybe
I'll recover and watch this again someday", and even manually compare
the unused outputs of several disks to know which ones still have it.
(Well, they *are* still the first candidates to purge if I run out of
disk space...)

So it's become like a secondary git-annex inside git-annex. Oh come on.

But I'm currently trying to clean this up (as part of the migration off
virt2 – oh yeah, whoever though it was a good idea to store anime on
work's virtualization servers, regardless of the terabytes of free disk
space they've got) and oh shit I'm doing this thing again

(╯°□°)╯︵ ┻━┻