fix-dropbox: further notes
Of interest is that the Windows version of Dropbox uses standard Windows
DPAPI to obscure the hostkeys (which is the correct choice), but *also*
stores a plaintext copy of the same keys right besides the protected
version.  (That's what the "…-p" value in registry are.)

I should strace the Linux Dropbox startup process and see which files it
looks at. Perhaps it also supports disabling obfuscation. (Naturally the
Python bytecode itself will probably be obfuscated...)