dmenu_run: add a hack to bypass autocompletion
With current dmenu it is impossible to enter a command that's a subset
of an existing command in history, unless that exact command itself also
exists in history. For example, if 'foobar' is in history but 'foo' is
not, then it is impossible to run 'foo'.

This hack lets me input it as '+foo' instead.