mkca: use 2048-bit keys
In hardware tokens, 3072 is even less commonly supported than 4096.
We're *fine* with the security level of RSA-2048, even if it's only
approx. 112 bits of security.

(NIST says it's fine until 2030, and in our "tiny private CA that nobody
cares about" case it's fine even until 2040 or so, at which point these
hardware tokens will definitely be dead.)