rcons: re-do Dunelab
Reconnected via AUX and old cable again, because the new ones are acting
weird... but so is the old one, so it must be a Linux pl2303 driver bug.
Top is now connected via the old USB adapter to a NetBSD VM which seems
to handle it fine; the other two are on AUX lines for now.