experimentally add a .hostrun wrapper for compiled tools
This would avoid having to dynamically determine the obj/ subdirectory
when generating $PATH (though it is a pretty basic "host.$HOSTNAME"
expansion, but $HOSTNAME is not guaranteed to be set in non-bash cases).

Of course, really the entire obj/*/ structure is overengineered; no
hosts actually share the home directory over NFS or anything anymore.