setptr: manually look up the master server
It seems that current BIND stores zone data in a non-case-preserving
manner, so if other records (such as the PTRs in our RPZ zone) are in
upper case, then the same tail node will be reused for the SOA MNAME --
which then results in nsupdate looking for the wrong Kerberos principal.
Do the lookup manually.